requirements.txt
aniso8601==8.0.0
astroid==2.4.2
attrs==20.2.0
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.2.1
cx-Oracle==8.0.1
elasticsearch==7.10.0
Flask==1.1.2
Flask-Cors==3.0.9
flask-restplus==0.13.0
flask-restx==0.2.0
Flask-Script==2.0.6
idna==2.10
isort==5.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.21
protobuf==3.13.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pyOpenSSL==20.0.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rope==0.18.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
XlsxWriter==1.3.7
